fix declaration of strtoull
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17864 06827809-a52a-0410-b366-d66718629ded

+++ include/net-snmp/library/system.h
@@ -137,7 +137,7 @@ SOFTWARE.
     unsigned long   strtoul(const char *, char **, int);
 #endif
 #ifndef HAVE_STRTOULL
-    unsigned long   strtoull(const char *, char **, int);
+    unsigned long long strtoull(const char *, char **, int);
 #endif
 #ifndef HAVE_STRTOK_R
     char           *strtok_r(char *, const char *, char **);
